refactor: remove redundant components
