--- src/components/cv-modal/cv-button/CVButton.tsx
@@ -1,15 +1,0 @@
-import { CVButtonStyle } from './CVButtonStyle';
-import { ButtonProps } from 'antd';
-
-function CVButton({ onClick }: ButtonProps) {
-  return (
-    <CVButtonStyle
-      type='primary'
-      description="Показати прев'ю"
-      shape='circle'
-      onClick={onClick}
-    />
-  );
-}
-
-export default CVButton;
