+++ src/components/cv/CVStyle.ts
@@ -1,4 +1,9 @@
 import styled from 'styled-components';
+import { Button } from 'antd';
+import { Typography } from 'antd';
+import colors from '../../colors';
+
+const { Paragraph } = Typography;
 
 export const CVStyle = styled.div`
   display: flex;
@@ -7,9 +12,35 @@ export const CVStyle = styled.div`
   & .ant-typography {
     margin: 0;
   }
-  @media print {
-    @page {
-      size: 100% 100%;
-    }
-  }
+`;
+
+export const DownloadButtonStyle = styled(Button)`
+  position: absolute;
+  top: -55px;
+  left: 50%;
+  transform: translateX(-50%);
+`;
+
+export const LeftBarStyle = styled.div`
+  padding: 20px;
+  width: 30%;
+  background-color: #f4f9f9;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+`;
+
+export const MainContentStyle = styled.div`
+  padding: 20px;
+  width: 70%;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+`;
+
+export const PositionStyle = styled(Paragraph)`
+  color: ${colors.dark};
+  padding-bottom: 10px;
+  border-bottom: 2px solid ${colors.dark};
+  font-size: 16px;
 `;
