--- src/components/cv/download-button/DownloadButton.tsx
@@ -1,8 +1,0 @@
-import { ButtonProps } from 'antd';
-import { DownloadButtonStyle } from './DownloadButtonStyle';
-
-function DownloadButton({ onClick }: ButtonProps) {
-  return <DownloadButtonStyle onClick={onClick}>Скачати</DownloadButtonStyle>;
-}
-
-export default DownloadButton;
