--- src/components/cv/download-button/DownloadButtonStyle.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import { Button } from 'antd';
-
-export const DownloadButtonStyle = styled(Button)`
-  position: absolute;
-  top: -55px;
-  left: 50%;
-  transform: translateX(-50%);
-`;
