--- src/components/cv/left-bar/LeftBar.tsx
@@ -1,11 +1,0 @@
-import { LeftBarStyle } from './LeftBarStyle';
-
-interface LeftBarProps {
-  children?: React.ReactNode;
-}
-
-function LeftBar({ children }: LeftBarProps) {
-  return <LeftBarStyle>{children}</LeftBarStyle>;
-}
-
-export default LeftBar;
