--- src/components/cv/left-bar/LeftBarStyle.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const LeftBarStyle = styled.div`
-  padding: 20px;
-  width: 30%;
-  background-color: #f4f9f9;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-`;
