--- src/components/cv/main-content/MainContent.tsx
@@ -1,11 +1,0 @@
-import { MainContentStyle } from './MainContentStyle';
-
-interface MainContentProps {
-  children?: React.ReactNode;
-}
-
-function MainContent({ children }: MainContentProps) {
-  return <MainContentStyle>{children}</MainContentStyle>;
-}
-
-export default MainContent;
