--- src/components/cv/main-content/MainContentStyle.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const MainContentStyle = styled.div`
-  padding: 20px;
-  width: 70%;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-`;
