+++ src/components/cv/title/Title.tsx
@@ -1,4 +1,11 @@
-import { TitleStyle, TitleProps } from './TitleStyle';
+import { TitleStyle } from './TitleStyle';
+import { TitleProps as AntTitleProps } from 'antd/es/typography/Title';
+
+export type Size = 'small' | 'medium' | 'large';
+
+interface TitleProps extends AntTitleProps {
+  size: Size;
+}
 
 function Title({ children, size, level }: TitleProps) {
   return (
