+++ src/components/cv/title/TitleStyle.ts
@@ -1,13 +1,9 @@
 import styled from 'styled-components';
 import { Typography } from 'antd';
+import type { Size } from './Title';
 import colors from '../../../colors';
-import type { TitleProps as AntTitleProps } from 'antd/es/typography/Title';
 
 const { Title } = Typography;
-
-export interface TitleProps extends AntTitleProps {
-  size: 'small' | 'medium' | 'large';
-}
 
 const sizes = {
   small: '14px',
@@ -15,10 +11,9 @@ const sizes = {
   large: '25px',
 };
 
-export const TitleStyle = styled(Title)<TitleProps>`
+export const TitleStyle = styled(Title)<{ size: Size }>`
   &.ant-typography {
     font-size: ${(props) => sizes[props.size]};
-    font-size: '30px';
     color: ${colors.dark};
     text-transform: uppercase;
   }
